Filestore: Move snapshot related code into its own file.
License: MIT
