+++ filestore/datastore.go
@@ -31,23 +31,6 @@ const (
 	VerifyAlways
 )
 
-type readonly interface {
-	Get(key []byte, ro *opt.ReadOptions) (value []byte, err error)
-	Has(key []byte, ro *opt.ReadOptions) (ret bool, err error)
-	NewIterator(slice *util.Range, ro *opt.ReadOptions) iterator.Iterator
-}
-
-type Basic struct {
-	db readonly
-	ds *Datastore
-}
-
-type Snapshot struct {
-	*Basic
-}
-
-func (ss *Snapshot) Defined() bool { return ss.Basic != nil }
-
 type Datastore struct {
 	db     *leveldb.DB
 	verify VerifyWhen
@@ -76,36 +59,15 @@ type Datastore struct {
 	//maintLock  sync.RWMutex
 }
 
-func (b *Basic) Verify() VerifyWhen { return b.ds.verify }
-
-func (d *Basic) DB() readonly { return d.db }
-
-func (d *Datastore) DB() *leveldb.DB { return d.db }
-
-func (d *Datastore) AsBasic() *Basic { return &Basic{d.db, d} }
-
-func (d *Basic) AsFull() *Datastore { return d.ds }
-
-func (d *Datastore) GetSnapshot() (Snapshot, error) {
-	if d.snapshot.Defined() {
-		d.snapshotUsed = true
-		return d.snapshot, nil
-	}
-	ss, err := d.db.GetSnapshot()
-	if err != nil {
-		return Snapshot{}, err
-	}
-	return Snapshot{&Basic{ss, d}}, nil
+type readonly interface {
+	Get(key []byte, ro *opt.ReadOptions) (value []byte, err error)
+	Has(key []byte, ro *opt.ReadOptions) (ret bool, err error)
+	NewIterator(slice *util.Range, ro *opt.ReadOptions) iterator.Iterator
 }
 
-func (d *Datastore) releaseSnapshot() {
-	if !d.snapshot.Defined() {
-		return
-	}
-	if !d.snapshotUsed {
-		d.snapshot.db.(*leveldb.Snapshot).Release()
-	}
-	d.snapshot = Snapshot{}
+type Basic struct {
+	db readonly
+	ds *Datastore
 }
 
 func (d *Datastore) updateOnGet() bool {
@@ -501,45 +463,4 @@ func (d *Datastore) Close() error {
 
 func (d *Datastore) Batch() (ds.Batch, error) {
 	return ds.NewBasicBatch(d), nil
-}
-
-func NoOpLocker() sync.Locker {
-	return noopLocker{}
-}
-
-type noopLocker struct{}
-
-func (l noopLocker) Lock() {}
-
-func (l noopLocker) Unlock() {}
-
-type addLocker struct {
-	adders int
-	lock   sync.Mutex
-	ds     *Datastore
-}
-
-func (l *addLocker) Lock() {
-	l.lock.Lock()
-	defer l.lock.Unlock()
-	if l.adders == 0 {
-		l.ds.releaseSnapshot()
-		l.ds.snapshot, _ = l.ds.GetSnapshot()
-	}
-	l.adders += 1
-	log.Debugf("acquired add-lock refcnt now %d\n", l.adders)
-}
-
-func (l *addLocker) Unlock() {
-	l.lock.Lock()
-	defer l.lock.Unlock()
-	l.adders -= 1
-	if l.adders == 0 {
-		l.ds.releaseSnapshot()
-	}
-	log.Debugf("released add-lock refcnt now %d\n", l.adders)
-}
-
-func (d *Datastore) AddLocker() sync.Locker {
-	return &d.addLocker
 }
